+++ metadata-ingestion/src/datahub/ingestion/source/gcs/gcs_source.py
@@ -15,6 +15,10 @@ from datahub.ingestion.api.decorators import (
     support_status,
 )
 from datahub.ingestion.api.source import Source, SourceCapability, SourceReport
+from datahub.ingestion.api.source_helpers import (
+    auto_status_aspect,
+    auto_workunit_reporter,
+)
 from datahub.ingestion.api.workunit import MetadataWorkUnit
 from datahub.ingestion.source.aws.aws_common import AwsConnectionConfig
 from datahub.ingestion.source.data_lake_common.config import PathSpecsConfigMixin
@@ -22,7 +26,6 @@ from datahub.ingestion.source.data_lake_common.data_lake_utils import PLATFORM_G
 from datahub.ingestion.source.data_lake_common.path_spec import PathSpec, is_gcs_uri
 from datahub.ingestion.source.s3.config import DataLakeSourceConfig
 from datahub.ingestion.source.s3.source import S3Source
-from datahub.utilities.source_helpers import auto_status_aspect, auto_workunit_reporter
 
 logger: logging.Logger = logging.getLogger(__name__)
