+++ metadata-ingestion/src/datahub/ingestion/source/metadata/business_glossary.py
@@ -20,14 +20,14 @@ from datahub.ingestion.api.decorators import (
     support_status,
 )
 from datahub.ingestion.api.source import Source, SourceReport
-from datahub.ingestion.api.workunit import MetadataWorkUnit
-from datahub.ingestion.graph.client import DataHubGraph
-from datahub.utilities.registries.domain_registry import DomainRegistry
-from datahub.utilities.source_helpers import (
+from datahub.ingestion.api.source_helpers import (
     auto_status_aspect,
     auto_workunit,
     auto_workunit_reporter,
 )
+from datahub.ingestion.api.workunit import MetadataWorkUnit
+from datahub.ingestion.graph.client import DataHubGraph
+from datahub.utilities.registries.domain_registry import DomainRegistry
 from datahub.utilities.urn_encoder import UrnEncoder
 
 logger = logging.getLogger(__name__)
